==> text_classifier/01 process_data.ipynb <==
import json 
import pandas as pd 

from tqdm.autonotebook import tqdm

# 对Yelps数据的处理

## 查看原始数据

with open("../data/review.json", "r", encoding="utf8") as f:
    for line in f:
        text = json.loads(line)
        #print(text)
        temp = pd.DataFrame(text, index=[0])
        print(temp.head())
        break

temp.head()

## 对数据进行转换

with open("../data/review.json", "r", encoding="utf8") as f:
    for i, line in tqdm(enumerate(f)):
        record = json.loads(line)
        record = pd.DataFrame(record, index=[0])
        if i == 0: 
            record.to_csv("../data/review.csv", header=True, encoding="utf-8", index=False)
        else:
            record.to_csv("../data/review.csv", header=False, encoding="utf-8", mode="a", index=False)

data = pd.read_csv("../data/review.csv", nrows=10000)

data.head()

## 将数据集转换成训练需要的格式

# 对标签进行处理，大小3的表示pos，小于等于3的表示neg
data['emotion'] = data['stars'].apply(lambda x: int(x>3))

import nltk
from nltk import tokenize
from nltk.corpus import stopwords
import spacy
import re 
from nltk.stem import WordNetLemmatizer

# 划分训练集和验证集
from sklearn.model_selection import train_test_split

stoplists = stopwords.words("english")

wnl = WordNetLemmatizer()

# 添加需要去除的标点符号集，问号和感叹号除外
puncts = [',', '.', '"', ':', ')', '(', '-', '?', '!', '|', ';', "'", '$', '&', '/', 
          '[', ']', '>', '%', '=', '#', '*', '+', '\\', '•',  '~', '@', '£', '·', '_', 
          '{', '}', '©', '^', '®', '`',  '<', '→', '°', '€', '™', '›',  '♥', '←', '×', 
          '§', '″', '′', 'Â', '█', '½', 'à', '…', '“', '★', '”', '–', '●', 'â', '►', 
          '−', '¢', '²', '¬', '░', '¶', '↑', '±', '¿', '▾', '═', '¦', '║', '―', '¥',
          '▓', '—', '‹', '─', '▒', '：', '¼', '⊕', '▼', '▪', '†', '■', '’', '▀', '¨', 
          '▄', '♫', '☆', 'é', '¯', '♦', '¤', '▲', 'è', '¸', '¾', 'Ã', '⋅', '‘', '∞', 
          '∙', '）', '↓', '、', '│', '（', '»', '，', '♪', '╩', '╚', '³', '・', '╦', '╣', 
          '╔', '╗', '▬', '❤', 'ï', 'Ø', '¹', '≤', '‡', '√', ]

# 定义一些常见的缩写
contraction_patterns = [ (r'won\'t', 'will not'), (r'can\'t', 'cannot'), (r'i\'m', 'i am'),
                        (r'isn\'t', 'is not'), (r'(\w+)\'ll', '\g<1> will'), 
                        (r'(\w+)n\'t', '\g<1> not'),(r'(\w+)\'ve', '\g<1> have'), 
                        (r'(\w+)\'s', '\g<1> is'), (r'(\w+)\'re', '\g<1> are'), 
                        (r'(\w+)\'d', '\g<1> would'), (r'&', 'and'), (r'dammit', 'damn it'),
                        (r'dont', 'do not'), (r'wont', 'will not') ]

def clean_text(text):
    # 去除对情感分类没有用的数字
    text = re.sub("[0-9]+", "", text)
    # 对重复出现的标点进行天魂
    text = re.sub(r"(\!)\1+", "multiExclamation", text)
    text = re.sub(r"(\?)\1+", "multiQuestion", text)
    text = re.sub(r"(\.)\1+", "multiStop", text)
    
    # 对缩写进行替换
    patterns = [(re.compile(regex), repl) for (regex, repl) in contraction_patterns]
    for (pattern, repl) in patterns:
        (text, count) = re.subn(pattern, repl, text)
    
    # 在标点前面加上空格
    for punct in puncts:
        text = text.replace(punct, f" {punct} ")
    
    #print(text)
    # 对文本进行分词
    text_split = tokenize.word_tokenize(text)
    text = [word for word in text_split if word not in stoplists]
    text = [wnl.lemmatize(word) for word in text]
    
    return " ".join(text)

data['text'] = data['text'].apply(clean_text）

X_train, X_test, y_train, y_test = train_test_split(data['text'].values, data['emotion'].values,
                                                   test_size=0.3, random_state=2019, shuffle=True,
                                                   stratify=data['emotion'].values)

# 将训练集和测试集写入文件中
with open("../data/yelps_train_fasttext.txt", "w", encoding="utf8") as f: 
    train = zip(X_train, y_train)
    for X, y in tqdm(train):
        label = "pos" if y > 0 else "neg"
        record = X + "\t__label__"+label+"\n"
        f.write(record)

# 将测试集写入文件中
with open("../data/yelps_test_fasttext.txt", "w", encoding="utf-8") as f: 
    test = zip(X_test, y_test)
    for X, y in tqdm(test):
        label = "pos" if y>0 else "neg"
        record = X+"\t__label__"+label+"\n"
        f.write(record)

# 对IMDB review的处理

import pandas as pd 
from bs4 import BeautifulSoup

unlabel = pd.read_csv("../data/rawData/unlabeledTrainData.tsv", sep="\t", lineterminator="\n", 
                     error_bad_lines=False)

unlabel.head()

label = pd.read_csv("../data/rawData/labeledTrainData.tsv", sep="\t", lineterminator="\n",
                   error_bad_lines=False)

label.head()

def get_rate(string):
    string = string.split("_")
    return int(string[1])

label["rate"] = label["id"].apply(get_rate)

label.head()

def clean_text(text):
    # 添加需要去除的标点符号集，问号和感叹号除外
    puncts = [',', '.', '"', ':', ')', '(', '-', '?', '!', '|', ';', "'", '$', '&', '/', 
              '[', ']', '>', '%', '=', '#', '*', '+', '\\', '•',  '~', '@', '£', '·', '_', 
              '{', '}', '©', '^', '®', '`',  '<', '→', '°', '€', '™', '›',  '♥', '←', '×', 
              '§', '″', '′', 'Â', '█', '½', 'à', '…', '“', '★', '”', '–', '●', 'â', '►', 
              '−', '¢', '²', '¬', '░', '¶', '↑', '±', '¿', '▾', '═', '¦', '║', '―', '¥',
              '▓', '—', '‹', '─', '▒', '：', '¼', '⊕', '▼', '▪', '†', '■', '’', '▀', '¨', 
              '▄', '♫', '☆', 'é', '¯', '♦', '¤', '▲', 'è', '¸', '¾', 'Ã', '⋅', '‘', '∞', 
              '∙', '）', '↓', '、', '│', '（', '»', '，', '♪', '╩', '╚', '³', '・', '╦', '╣', 
              '╔', '╗', '▬', '❤', 'ï', 'Ø', '¹', '≤', '‡', '√', ]

    # 定义一些常见的缩写
    contraction_patterns = [ (r'won\'t', 'will not'), (r'can\'t', 'cannot'), (r'i\'m', 'i am'),
                            (r'isn\'t', 'is not'), (r'(\w+)\'ll', '\g<1> will'), 
                            (r'(\w+)n\'t', '\g<1> not'),(r'(\w+)\'ve', '\g<1> have'), 
                            (r'(\w+)\'s', '\g<1> is'), (r'(\w+)\'re', '\g<1> are'), 
                            (r'(\w+)\'d', '\g<1> would'), (r'&', 'and'), (r'dammit', 'damn it'),
                            (r'dont', 'do not'), (r'wont', 'will not') ]    
    # 去除对情感分类没有用的数字
    text = re.sub("[0-9]+", "", text)
    # 对重复出现的标点进行天魂
    text = re.sub(r"(\!)\1+", "multiExclamation", text)
    text = re.sub(r"(\?)\1+", "multiQuestion", text)
    text = re.sub(r"(\.)\1+", "multiStop", text)
    
    # 对缩写进行替换
    patterns = [(re.compile(regex), repl) for (regex, repl) in contraction_patterns]
    for (pattern, repl) in patterns:
        (text, count) = re.subn(pattern, repl, text)
    
    # 在标点前面加上空格
    for punct in puncts:
        text = text.replace(punct, f" {punct} ")
    
    #print(text)
    # 对文本进行分词
    text_split = tokenize.word_tokenize(text)
    text = [word for word in text_split if word not in stoplists]
    text = [wnl.lemmatize(word) for word in text]
    
    return " ".join(text)

unlabel.shape

label.shape

unlabel['review'] = unlabel['review'].apply(clean_text)

label['review'] = label['review'].apply(clean_text)

label.head()

def cleanReview(string):
    beau = BeautifulSoup(string)
    newstring = beau.get_text()
    newstring = newstring.strip().lower()
    return newstring

unlabel['review'] = unlabel["review"].apply(cleanReview)

label['review'] = label['review'].apply(cleanReview)

unlabel.head()

label.head()

# 过滤HTML中的标签
# 将HTML中标签等信息去掉
# @param htmlstr HTML字符串
def filter_tags(htmlstr):
    # 先过滤CDATA
    # re_cdata = re.compile('<!\[CDATA\[[^>]*//\]\]>', re.I)  # 匹配CDATA
    re_cdata = re.compile('(<!\[CDATA\[)[\s\S]*?(//\]\]>)', re.I)  # 匹配CDATA

    # re_script = re.compile('<\s*script[^>]*>[^<]*<\s*/\s*script\s*>', re.I)  # Script
    re_script = re.compile(
        '(<\s*script\s*[^>]*>)[\s\S]*?(</\s*script\s*>)', re.I)  # Script

    re_style = re.compile(
        '<\s*style[^>]*>[^<]*<\s*/\s*style\s*>', re.I)  # style
    re_br = re.compile('<br\s*?/?>')  # 处理换行
    re_h = re.compile('</?\w+[^>]*>')  # HTML标签
    re_comment = re.compile('<!--[^>]*-->')  # HTML注释
    re_htmlhead = re.compile('<!DOCTYPE html[^>]*>')  # HTML头部

    s = re_cdata.sub('', htmlstr)  # 去掉CDATA
    s = re_script.sub('', s)  # 去掉SCRIPT
    s = re_style.sub('', s)  # 去掉style
    s = re_br.sub('\n', s)  # 将br转换为换行
    s = re_h.sub('', s)  # 去掉HTML 标签
    s = re_comment.sub('', s)  # 去掉HTML注释

    s = re_htmlhead.sub('', s)  # HTML头部

    # 过滤连续空格
    space_line = re.compile(' +')
    s = space_line.sub(' ', s)

    # 去掉连续的空行
    blank_line = re.compile('\n+')
    s = blank_line.sub('\n', s)

    # 过滤只有空格的行
    line = re.compile(' \n')
    s = line.sub('', s)

    s = replaceCharEntity(s)  # 替换实体
    
    s = re.sub(r"</? \w+[^>]*>", "", s)
    
    
    return s


# 替换常用HTML字符实体.
# 使用正常的字符替换HTML中特殊的字符实体.
# 你可以添加新的实体字符到CHAR_ENTITIES中,处理更多HTML字符实体.
# @param htmlstr HTML字符串.
def replaceCharEntity(htmlstr):
    CHAR_ENTITIES = {'nbsp': ' ', '160': ' ',
                     'lt': '<', '60': '<',
                     'gt': '>', '62': '>',
                     'amp': '&', '38': '&',
                     'quot': '"', '34': '"', }

    re_charEntity = re.compile(r'&#?(?P<name>\w+);')
    sz = re_charEntity.search(htmlstr)
    while sz:
        # entity = sz.group()  # entity全称，如&gt;
        key = sz.group('name')  # 去除&;后entity,如&gt;为gt
        try:
            htmlstr = re_charEntity.sub(CHAR_ENTITIES[key], htmlstr, 1)
            sz = re_charEntity.search(htmlstr)
        except KeyError:
            # 以空串代替
            htmlstr = re_charEntity.sub('', htmlstr, 1)
            sz = re_charEntity.search(htmlstr)
    return htmlstr


def repalce(s, re_exp, repl_string):
    return re_exp.sub(repl_string, s)

unlabel['review'] = unlabel['review'].apply(filter_tags)

unlabel.head()

label["review"] = label['review'].apply(filter_tags)

# 将所有的文本输出用于训练词向量
new_df = pd.concat([unlabel['review'], label['review']], axis=0)

new_df.to_csv("../data/preprocess/word2vec.txt", index=False)

# 将有标签的处理之后的数据保留
label = label[["review", "sentiment", "rate"]]
label.to_csv("../data/preprocess/labeldTrain.csv", index=False)

label.head()

label['sentiment'] = label['sentiment'].apply(lambda x: "__label__pos" if x > 0 else "__label__neg")

label.head()

X_train, X_test, y_train, y_test = train_test_split(label['review'].values, label['sentiment'].values,
                                                   test_size=0.25, stratify=label['sentiment'].values,
                                                   random_state=2019, shuffle=True)

with open("../data/preprocess/imdb_fasttext_train.txt", "w", encoding="utf8") as f: 
    train = zip(X_train, y_train)
    for x, y in tqdm(train):
        f.write(x + '\t' + y + '\n')

with open("../data/preprocess/imdb_fasttext_test.txt", "w", encoding="utf-8") as f: 
    test = zip(X_test, y_test)
    for x, y in tqdm(test):
        f.write(x+"\t"+y+"\n")
