==> text_classifier/02 fasttext.ipynb <==
import fasttext
import pandas as pd 
import numpy as np 
import gc 
import os 
import nltk

from collections import Counter

from tqdm.autonotebook import tqdm

# 读取数据并进行预处理

data = pd.read_csv("../data/review.csv")

data.head()

# 删除掉无用的列，只保留有用的列
data = data[['review_id', 'stars', 'text']]

# 对标签进行统计分析
Counter(data['stars'])

# 对标签进行处理，大小3的表示pos，小于等于3的表示neg
data['emotion'] = data['stars'].apply(lambda x: int(x>3))

data.head()

# 对文本进行处理

import nltk
from nltk import tokenize
from nltk.corpus import stopwords
import spacy
import re 
from nltk.stem import WordNetLemmatizer

stoplists = stopwords.words("english")

wnl = WordNetLemmatizer()

# 添加需要去除的标点符号集，问号和感叹号除外
puncts = [',', '.', '"', ':', ')', '(', '-', '?', '!', '|', ';', "'", '$', '&', '/', 
          '[', ']', '>', '%', '=', '#', '*', '+', '\\', '•',  '~', '@', '£', '·', '_', 
          '{', '}', '©', '^', '®', '`',  '<', '→', '°', '€', '™', '›',  '♥', '←', '×', 
          '§', '″', '′', 'Â', '█', '½', 'à', '…', '“', '★', '”', '–', '●', 'â', '►', 
          '−', '¢', '²', '¬', '░', '¶', '↑', '±', '¿', '▾', '═', '¦', '║', '―', '¥',
          '▓', '—', '‹', '─', '▒', '：', '¼', '⊕', '▼', '▪', '†', '■', '’', '▀', '¨', 
          '▄', '♫', '☆', 'é', '¯', '♦', '¤', '▲', 'è', '¸', '¾', 'Ã', '⋅', '‘', '∞', 
          '∙', '）', '↓', '、', '│', '（', '»', '，', '♪', '╩', '╚', '³', '・', '╦', '╣', 
          '╔', '╗', '▬', '❤', 'ï', 'Ø', '¹', '≤', '‡', '√', ]

# 定义一些常见的缩写
contraction_patterns = [ (r'won\'t', 'will not'), (r'can\'t', 'cannot'), (r'i\'m', 'i am'),
                        (r'isn\'t', 'is not'), (r'(\w+)\'ll', '\g<1> will'), 
                        (r'(\w+)n\'t', '\g<1> not'),(r'(\w+)\'ve', '\g<1> have'), 
                        (r'(\w+)\'s', '\g<1> is'), (r'(\w+)\'re', '\g<1> are'), 
                        (r'(\w+)\'d', '\g<1> would'), (r'&', 'and'), (r'dammit', 'damn it'),
                        (r'dont', 'do not'), (r'wont', 'will not') ]

def clean_text(text):
    # 去除对情感分类没有用的数字
    text = re.sub("[0-9]+", "", text)
    # 对重复出现的标点进行天魂
    text = re.sub(r"(\!)\1+", "multiExclamation", text)
    text = re.sub(r"(\?)\1+", "multiQuestion", text)
    text = re.sub(r"(\.)\1+", "multiStop", text)
    
    # 对缩写进行替换
    patterns = [(re.compile(regex), repl) for (regex, repl) in contraction_patterns]
    for (pattern, repl) in patterns:
        (text, count) = re.subn(pattern, repl, text)
    
    # 在标点前面加上空格
    for punct in puncts:
        text = text.replace(punct, f" {punct} ")
    
    #print(text)
    # 对文本进行分词
    text_split = tokenize.word_tokenize(text)
    text = [word for word in text_split if word not in stoplists]
    text = [wnl.lemmatize(word) for word in text]
    
    return " ".join(text)

clean_text("I'll a test file ! ... !@@##$()")

data['text'] = data['text'].apply(lambda s: clean_text(s))

data.head()

# 划分训练集和验证集
from sklearn.model_selection import train_test_split

X_train, X_test, y_train, y_test = train_test_split(data['text'].values, data['emotion'].values,
                                                   test_size=0.3, random_state=2019, shuffle=True,
                                                   stratify=data['emotion'].values)

X_train[0]

下面将文本写成fasttext需要使用的文本格式，比如：

Big Juicy Gourmet Burger melted cheese bacon crumbles underneath . &emsp;&emsp;   \_\_label\_\_pos 

即每一行前面表示文本，后面表示标签

# 将训练集和测试集写入文件中
with open("../data/yelps_train_fasttext.txt", "w", encoding="utf8") as f: 
    train = zip(X_train, y_train)
    for X, y in tqdm(train):
        label = "pos" if y > 0 else "neg"
        record = X + "\t__label__"+label+"\n"
        f.write(record)

# 将测试集写入文件中
with open("../data/yelps_test_fasttext.txt", "w", encoding="utf-8") as f: 
    test = zip(X_test, y_test)
    for X, y in tqdm(test):
        label = "pos" if y>0 else "neg"
        record = X+"\t__label__"+label+"\n"
        f.write(record)

X_test.shape

# 训练fastText模型

## 对yelps数据集进行分类

classifier = fasttext.train_supervised("../data/yelps_train_fasttext.txt", epoch=25, lr=0.25, wordNgrams=2,
                                      verbose=2, minCount=2)

# 测试训练的结果
result = classifier.test("../data/yelps_test_fasttext.txt")

# 第一个参数表示测试集大小，第二个参数表示准确率，第三个参数表示召回率
result

- 共测试了715137条评论

- <font color=red size=4>**得到 P为90.338%，R为90.338%**</font>

# 预测结果第一个参数表示标签，第二个参数表示概率
classifier.predict(["This is bad !"])

## 对IMDB数据集进行分类

classifier = fasttext.train_supervised("../data/preprocess/imdb_fasttext_train.txt", epoch=25, lr=0.25,
                                      wordNgrams=2, verbose=2, minCount=2)

# 测试集结果
result = classifier.test("../data/preprocess/imdb_fasttext_test.txt")

result

- 共测试了6250条评论

- <font color=red size=4>**得到P为88.8%，R为88.8%**</font>

# 用tensorflow实现fastText

import tensorflow as tf
import numpy as np 
import os 

class fastTextModel():
    def __init__(self, sequence_length, num_classes, vocab_size, embedding_size, 
                learning_rate, decay_steps, decay_rate, l2_reg_lambda, is_training=True, 
                initializer=tf.random_normal_initializer(stddev=0.1)):
        self.vocab_size = vocab_size
        self.embedding_size = embedding_size
        self.num_classes = num_classes
        self.sequence_length = sequence_length
        self.learning_rate = learning_rate
        self.decay_steps = decay_steps
        self.decay_rate = decay_rate
        self.is_training = is_training
        self.l2_reg_lambda = l2_reg_lambda
        self.initializer = initializer
        
        self.input_x = tf.placeholder(tf.int32, [None, self.sequence_length], name="input_x")
        self.input_y = tf.placeholder(tf.int32, [None, self.num_classes], name="input_y")
        
        self.global_step = tf.Variable(0, trainable=False, name="global_step")
        self.instantiate_weight()
        self.logits = self.inference()
        self.loss_val = self.loss()
        self.train_op = self.train() 
        
        self.predictions = tf.argmax(self.logits, axis=1, name="predictions")
        correct_predictions = tf.equal(self.predictions, tf.argmax(self.input_y, 1))
        self.accuracy = tf.reduce_mean(tf.cast(correct_predictions, tf.float32), name="accuracy")
        
    def instantiate_weight(self):
        with tf.name_scope("weights"):
            self.Embedding = tf.get_variable("Embedding", shape=[self.vocab_size, self.embedding_size],
                                            initializer=self.initializer)
            self.W_projection = tf.get_variable("W_projection", shape=[self.embedding_size, self.num_classes],
                                               initializer=self.initializer)
            self.b_projection = tf.get_variable("b_projection", shape=[self.num_classes])
            
    def inference(self):
        '''
        1. word embedding
        2. average embedding
        3. linear classifier
        '''
        # embedding layer
        with tf.name_scope("embedding"):
            words_embedding = tf.nn.embedding_lookup(self.Embedding, self.input_x)
            self.average_embedding = tf.reduce_mean(words_embedding, axis=1)  # 代表了这句话的embedding
        
        logits = tf.matmul(self.average_embedding, self.W_projection) + self.b_projection
        return logits
    
    def loss(self):
        with tf.name_scope("loss"):
            losses = tf.nn.softmax_cross_entropy_with_logits_v2(labels=self.input_y, logits=self.logits)
            data_loss = tf.reduce_mean(losses)
            l2_loss = tf.add_n([tf.nn.l2_loss(cand_var)] for cand_var in tf.trainable_variables() 
                              if "bias" not in cand_var.name) * self.l2_reg_lambda
            data_loss += l2_loss
            return data_loss
        
    def train(self):
        with tf.name_scope("train"):
            learning_rate = tf.train.exponential_decay(self.learning_rate, self.global_step, 
                                                      self.decay_steps, self.decay_rate,
                                                      staircase=True)
            train_op = tf.contrib.layers.optimize_loss(self.loss_val, 
                                                      global_step=self.global_step,
                                                      learning_rate=learning_rate,
                                                      optimizer="Adam")
            
            return train_op
