==> text_classifier/03 train_word2vec.ipynb <==
import logging
import gensim
from gensim.models.word2vec import Word2Vec, LineSentence

import pandas as pd 

# 训练yelps的词向量

yelps_train = pd.read_csv("../data/yelps/yelps_train_fasttext.txt", sep="\t", lineterminator="\n", header=None,
                         names=["review", "emotion"])

yelps_train.head()

yelps_train['emotion'] = yelps_train['emotion'].apply(lambda s: 1 if s[-3:]=="pos" else 0)

yelps_train.head()

yelps_test = pd.read_csv("../data/yelps/yelps_test_fasttext.txt", sep="\t", lineterminator="\n", header=None, 
                        names=["review", "emotion"])

yelps_test.head()

yelps_test['emotion'] = yelps_test['emotion'].apply(lambda s: 1 if s[-3:]=="pos" else 0)

yelps_test.head()

# 保存成训练word2vec的句子
new_df = pd.concat([yelps_train['review'], yelps_test['review']], axis=0)

new_df.to_csv("../data/yelps/word2vec.txt", header=False, index=False)

yelps_train.to_csv("../data/yelps/yelps_train.csv", index=False)

yelps_test.to_csv("../data/yelps/yelps_test.csv", index=False)

sentences = LineSentence("../data/yelps/word2vec.txt")

model = Word2Vec(sentences, size=200, sg=1, iter=10)

model.wv.save_word2vec_format("../data/yelps/wordvector.bin", binary=True)

# 训练reviews的word2vec模型

sentences = LineSentence("../data/imdb/word2vec.txt")

model = Word2Vec(sentences, size=200, sg=1, iter=10)

model.wv.save_word2vec_format("../data/imdb/wordvector.bin", binary=True)
