==> text_classifier/05 charCNN.ipynb <==
import os 
import time
import datetime
import json

import math 
import warnings
warnings.filterwarnings("ignore")

import numpy as np 
import pandas as pd 
import tensorflow as tf 

from sklearn.model_selection import train_test_split

from tqdm.autonotebook import tqdm

from utils import *

# 定义参数配置

class Config(dict):
    def __init__(self, path=None):
        super().__init__()
        ## 定义训练参数
        self['num_epochs'] = 2 
        self['evaluateEvery'] = 100 
        self['checkpointEvery'] = 100 
        self['learningRate'] = 0.001 
        
        ## 定义模型参数
        self['embeddingSize'] = 200 
        self['dropoutProb'] = 0.5 
        self['l2RegLambda'] = 0.0 
        
        ## 定义基础参数
        self['sequenceLength'] = 1014 
        self['batch_size'] = 128 
        self['dataSource'] = path   # 表示训练集文件所在的目录，csv文件，包含2-3列，类别列为`emotion`或者`sentiment`
        self['stopWordSource'] = "../data/english"  # 表示停止词所在的文件
        self['numClasses'] = 1  
        self['train_size'] = 0.8   # 训练集和测试集比例
        self.threshold = 0.5 
        
        ## 保存模型参数
        self['checkpoint_dir'] = "../model/charCNN/imdb/checkpoint"
        self['summary_dir'] = "../model/charCNN/imdb/summary"
        self['max_to_keep'] = 5
        
        
        ## 专用于charCNN的一些参数
        ### 3个参数分别表示 卷积核的数量，卷积核的高度，池化的尺寸
        self['convLayers'] = [[256, 7, 4], 
                              [256, 7, 4], 
                              [256, 3, 4]]
        self['fcLayers'] = [512]
        self['epsilon'] = 1e-3  # BN层中防止分母为0而加入的极小值
        self['decay'] = 0.999   # BN层中用来计算滑动平均的值 
        self['alphabet'] = "abcdefghijklmnopqrstuvwxyz0123456789-,;.!?:'\"/\\|_@#$%^*()~`+-=<>{}[]"
        

## 定义数据生成类

class Dataset(object):
    def __init__(self, config):
        self._dataSource = config['dataSource']
        self._sequenceLength = config['sequenceLength']
        self._trainSize = config['train_size']
        
        self.trainReviews = []
        self.trainLabels = []
        self.evalReviews = []
        self.evalLabels = []
        
        self._alphabet = config['alphabet']
        self.charEmbedding = None
        self._charToIndex = {}
        self._indexToChar = {}
        
    def _readData(self, filePath):
        '''
        从csv文件中读取数据
        '''
        df = pd.read_csv(filePath)
        if "sentiment" in df.columns:
            labels = df["sentiment"].tolist()
        elif "emotion" in df.columns:
            labels = df["emotion"].tolist()
        
        self.labelList = list(set(labels))
        
        review = df["review"].tolist()
        reviews = [[char for char in line if char != " "] for line in review]
        return reviews, labels 
    
    def _reviewProcess(self, review, sequenceLength, charToIndex):
        '''
        将数据集中的每条评论用index表示
        wordToIndex中的"pad"对应index为0
        '''
        reviewVec = np.zeros((sequenceLength))
        sequenceLen = sequenceLength
        
        # 判断当前的序列是否小于定义的固定序列长度
        if len(review) < sequenceLength:
            sequenceLen = len(review)
        for i in range(sequenceLen):
            if review[i] in charToIndex:
                reviewVec[i] = charToIndex[review[i]]
            else:
                reviewVec[i] = charToIndex["UNK"]
        
        return reviewVec
    
    def _genVocabulary(self, reviews, prefix=""):
        '''
        生成字符向量和字符-索引映射字典
        '''
        chars = [char for char in self._alphabet]
        vocab, charEmbedding = self._getcharEmbedding(chars)
        self.charEmbedding = charEmbedding
        
        self._charToIndex = dict(zip(vocab, list(range(len(vocab)))))
        self._indexToChar = dict(zip(list(range(len(vocab))), vocab))
        
        ## 将词汇-索引映射表保存为json数据
        basepath = "../data/charJson"
        if not os.path.exists(basepath):
            os.mkdir(basepath)
        with open(os.path.join(basepath, prefix+"_charToIndex.json"), "w", encoding="utf-8") as f: 
            json.dump(self._charToIndex, f)
        with open(os.path.join(basepath, prefix+"_indexToChar.json"), "w", encoding="utf-8") as f: 
            json.dump(self._indexToChar, f)
    
    def _getcharEmbedding(self, chars):
        '''
        将字符映射为one-hot向量
        '''
        alphabet = ["UNK"] + chars
        vocab = ["PAD"] + alphabet
        
        charEmbedding = []
        charEmbedding.append(np.zeros(len(alphabet), dtype="float32"))
        
        for i, alpha in enumerate(alphabet):
            onehot = np.zeros(len(alphabet), dtype="float32")
            # 生成每个字符对应的向量
            onehot[i] = 1 
            # 生成字符嵌入的向量矩阵
            charEmbedding.append(onehot)
        
        return vocab, np.array(charEmbedding)
    
    def _genTrainEvalData(self, x, y):
        '''
        生成训练集和验证集
        '''
        reviews = []
        labels = []
        
        # 遍历所有文本，将文本中的词转换为Indexbiaoshi 
        for i in range(len(x)):
            reviewVec = self._reviewProcess(x[i], self._sequenceLength, self._charToIndex)
            reviews.append(reviewVec)
            labels.append([y[i]])
        
        #print(len(y))
            
        trainReviews, evalReviews, trainLabels, evalLabels = train_test_split(reviews, labels,
                                                                             train_size=self._trainSize,
                                                                             shuffle=True,
                                                                             random_state=2019,
                                                                             stratify=y)
        return trainReviews, trainLabels, evalReviews, evalLabels
    
    def dataGen(self, path, prefix=""):
        '''
        初始化训练集和验证集
        '''
        # 初始化数据集
        reviews, labels = self._readData(self._dataSource)
        # 初始化词汇-索引映射表和词向量矩阵
        self._genVocabulary(reviews, prefix)
        # 初始化训练集和测试集
        trainReviews, trainLabels, evalReviews, evalLabels = self._genTrainEvalData(reviews, labels)
        
        self.trainReviews = np.array(trainReviews)
        self.trainLabels = np.array(trainLabels)
        self.evalReviews = np.array(evalReviews)
        self.evalLabels = np.array(evalLabels)

import numpy as np

class DataGenerator:
    def __init__(self, x, y):
        self.x = x
        self.y = y
        self.length = len(y)

    def next_batch(self, batch_size):
        '''
        生成每一个batch的数据集
        '''
        idx = np.random.choice(self.length, batch_size)
        #print(idx)
        #print(self.x[[1, 10, 9]])
        yield self.x[idx], self.y[idx]

    def iter_all(self, batch_size):
        '''
        按照batch迭代所有数据
        '''
        numBatches = self.length // batch_size
        for i in range(numBatches):
            start = i * batch_size
            end = start + batch_size
            batchX = np.array(self.x[start:end], dtype='int64')
            batchY = np.array(self.y[start:end], dtype="float32")
            yield batchX, batchY

# 定义charCNN模型和训练类

class CharCNN(BaseModel):
    def __init__(self, config, charEmbedding):
        super(CharCNN, self).__init__(config)
        self.charEmbedding = charEmbedding
        self.build_model()
        self.init_saver()
        
    def build_model(self):
        # 定义模型的输入
        self.inputX = tf.placeholder(tf.int32, [None, self.config['sequenceLength']], name="inputX")
        self.inputY = tf.placeholder(tf.float32, [None, 1], name="inputY")
        self.dropoutProb = tf.placeholder(tf.float32, name="dropoutProb")
        self.isTraining = tf.placeholder(tf.bool, name="isTraining")
        
        self.epsilon = self.config['epsilon']
        self.decay = self.config['decay']
        
        # 字符嵌入
        with tf.name_scope("embedding"):
            # 利用one-hot字符作为初始化词嵌入矩阵
            self.W = tf.Variable(tf.cast(self.charEmbedding, dtype=tf.float32, name="charEmbedding"),
                                name="W")
            # 获得字符嵌入
            self.embededChars = tf.nn.embedding_lookup(self.W, self.inputX)
            # 添加一个通道维度
            self.embededCharsExpand = tf.expand_dims(self.embededChars, -1)
            
        for i, cl in enumerate(self.config["convLayers"]):
            print("开始第", (i+1), "卷积层的处理")
            # 利用命名空间实现变量名复用
            with tf.name_scope(f"convLayer-{i+1}"):
                # 获取字符向量长度
                filterWidth = self.embededCharsExpand.get_shape()[2].value
                # filtershape=[height, width, in_channels, out_channels]
                filterShape = [cl[1], filterWidth, 1, cl[0]]
                
                stdv = 1 / math.sqrt(cl[0] * cl[1])
                # 初始化W和b的值
                wConv = tf.Variable(tf.random_uniform(filterShape, minval=-stdv, 
                                                     maxval=stdv),
                                   dtype=tf.float32, name="wConv")
                bConv = tf.Variable(tf.random_uniform(shape=[cl[0]], minval=-stdv,
                                                     maxval=stdv),
                                   name="bConv")
                
                conv = tf.nn.conv2d(self.embededCharsExpand, wConv, strides=[1,1,1,1],
                                   padding="VALID", name="Conv")
                # 加上偏差
                hConv = tf.nn.bias_add(conv, bConv)
                # 加上relu
                ## 得到的输出为[batch, seq_len-filter_size+1, 1, out_channels]
                hConv = tf.nn.relu(hConv)
                
                ## 卷积层中使用BN层
                #hConv = tf.layers.batch_normalization(hConv, training=self.isTraining)
                
                if cl[-1] is not None: 
                    # 维度是[batch, height, width, channel]
                    ksizeShape = [1, cl[2], 1, 1]
                    hPool = tf.nn.max_pool(hConv, ksize=ksizeShape, strides=ksizeShape,
                                          padding="VALID", name="pool")
                else:
                    hPool = hConv
                
                # 对维度进行转换，转换成卷积层的输入维度
                # 转换后的维度为 [batch, (seq_len-filter_size+1)/pool_size, channels, 1]
                self.embededCharsExpand = tf.transpose(hPool, [0, 1, 3, 2], name="transpose")
        
        with tf.name_scope("reshape"):
            ## 获取卷积层输出宽度
            fcDim = self.embededCharsExpand.get_shape()[1].value * self.embededCharsExpand.get_shape()[2].value
            self.inputReshape = tf.reshape(self.embededCharsExpand, [-1, fcDim])
        
        # 保存的是神经元的个数
        weights = [fcDim] + self.config['fcLayers']
        
        for i, fl in enumerate(self.config['fcLayers']):
            with tf.name_scope(f"fcLayer-{i+1}"):
                print("开始第", (i+1), "个全连接层的处理")
                stdv = 1 / math.sqrt(weights[i])
                
                # 定义全连接层初始化方法，均匀分布初始化w和b的值
                wFc = tf.Variable(tf.random_uniform([weights[i], fl], minval=-stdv, 
                                                   maxval=stdv),
                                 dtype=tf.float32, name="w")
                bFc = tf.Variable(tf.random_uniform(shape=[fl], minval=-stdv, 
                                                   maxval=stdv),
                                 dtype=tf.float32, name="b")
                self.fcInput = tf.nn.relu(tf.matmul(self.inputReshape, wFc) + bFc)
                ## 全连接之后使用batchNorm
                #self.fcInput = tf.layers.batch_normalization(self.fcInput, training=self.isTraining)
                
                with tf.name_scope("dropout"):
                    self.fcInputDrop = tf.nn.dropout(self.fcInput, rate=self.dropoutProb)
            self.inputReshape = self.fcInputDrop
        
        with tf.name_scope("outputLayer"):
            stdv = 1 / math.sqrt(weights[-1])
            self.logits = tf.layers.dense(self.inputReshape, self.config['numClasses'], name="dense",
                                              kernel_initializer=tf.random_uniform_initializer(
                                              minval=-stdv, maxval=stdv),
                                              bias_initializer=tf.random_uniform_initializer(
                                              minval=-stdv, maxval=stdv),
                                         activation=None)
            ## 进行二元分类
            self.predicitions = tf.cast(tf.greater_equal(self.predictions, 0.0),
                                      tf.float32, name="binaryPreds")
        
        with tf.name_scope("loss"):
            # 定义损失函数，对预测值进行softmax，再求交叉熵
            losses = tf.nn.sigmoid_cross_entropy_with_logits(logits=self.logits, labels=self.inputY)
            self.loss = tf.reduce_mean(losses)
            
        update_ops = tf.get_collection(tf.GraphKeys.UPDATE_OPS)
        with tf.control_dependencies(update_ops):
            self.train_op = tf.train.AdamOptimizer(
                self.config["learningRate"]).minimize(self.loss, global_step=self.global_step_tensor)
    
    def init_saver(self):
        '''
        初始化用于保存模型的对象
        '''
        self.saver = tf.train.Saver(max_to_keep=self.config['max_to_keep'])
    

## 定义训练的类

class Trainer(BaseTrain):
    def __init__(self, sess, model, data, config, logger):
        '''
        这里的data要求是元组形式，data[0]表示train对象，data[1]表示eval对象
        '''
        super(Trainer, self).__init__(sess, model, data, config, logger)
        self.train = data[0]
        self.eval = data[1]
        
    def train_epoch(self):
        num_iter_per_epoch = self.train.length // self.config["batch_size"]
        
        for _ in tqdm(range(num_iter_per_epoch)):
            loss, metrics, step = self.train_step()
            #print(metrics)
            train_acc = metrics["accuracy"]
            train_f_score = metrics["f_beta"]
            ## 将训练过程中的损失写入
            summaries_dict = {"loss": loss,
                             "acc": np.array(train_acc), 
                             "f_score": np.array(train_f_score)}
            self.logger.summarize(step, summarizer="train", scope="train_summary",
                                 summaries_dict=summaries_dict)
            
            if step % self.config['evaluateEvery'] == 0: 
                print("Train —— Step: {} | Loss: {} | Acc: {} | F1_Score: {}".format(
                    step, loss, train_acc, train_f_score))
                ## 对测试集进行评估
                print("\nEvaluation: \n")
                eval_losses = []
                eval_true = []
                eval_pred = []
                for batchEval in self.eval.iter_all(self.config["batch_size"]):
                    loss, predictions = self.eval_step(batchEval[0], batchEval[1])
                    eval_losses.append(loss)
                    eval_true.extend(batchEval[-1])
                    eval_pred.extend(predictions)
                
                getMetric = Metric(np.array(eval_pred), np.array(eval_true), self.config)
                metrics = getMetric.get_metrics()
                
                prec_mean = np.round(metrics['precision'], 5)
                recall_mean = np.round(metrics['recall'], 5)
                loss_mean = np.round(np.mean(eval_losses), 5)
                time_str = datetime.datetime.now().strftime("%Y-%m-%d %H:%M:%S %p")
                print("{} | Loss: {} | Precision: {} | Recall: {}".format(time_str, 
                                                                         loss_mean,
                                                                         prec_mean,
                                                                         recall_mean))
                summaries_dict = {"loss": np.array(loss_mean), 
                                 "precision": np.array(prec_mean),
                                 "recall": np.array(recall_mean)}
                self.logger.summarize(step, summarizer="test", scope="test_summary",
                                     summaries_dict=summaries_dict)
            if step % self.config["checkpointEvery"] == 0: 
                self.model.save(self.sess)
    
    def train_step(self):
        batch_x, batch_y = next(self.train.next_batch(self.config["batch_size"]))
        feed_dict = {self.model.inputX: batch_x, self.model.inputY: batch_y,
                    self.model.dropoutProb: self.config["dropoutProb"],
                    self.model.isTraining: True}
        
        _, loss, predictions, step = self.sess.run([self.model.train_op,
                                                   self.model.loss,
                                                   self.model.predictions, 
                                                   self.model.global_step_tensor],
                                                  feed_dict=feed_dict)
        #print(type(predictions))
        #print(type(batch_y))
        getMetric = Metric(predictions, batch_y, self.config)    
        metrics = getMetric.get_metrics()
            
        return loss, metrics, step 
    
    def eval_step(self, batch_x, batch_y):
        '''
        使用验证集进行测试
        '''
        #print(batch_x.shape, batch_y.shape)
        feed_dict = {self.model.inputX: batch_x, self.model.inputY: batch_y,
                    self.model.dropoutProb: 0.0, 
                    self.model.isTraining: False}
        loss, predictions = self.sess.run([self.model.loss, self.model.predictions],
                                         feed_dict=feed_dict)
            
        return loss, predictions

# 使用数据集进行训练和预测

## 使用IMDB数据集进行训练和预测

def main():
    # 实例化配置参数对象
    ## 执行训练数据文件名
    path = "../data/imdb/labeldTrain.csv"
    config = Config(path)
    
    create_dirs([config['summary_dir'], config["checkpoint_dir"]])
    
    data = Dataset(config)
    ## 生成训练集数据，第一个参数表示wordembedding文件所在的文件夹
    data.dataGen("../data/imdb/", prefix="imdb")
    train_X, train_y, eval_X, eval_y = data.trainReviews, data.trainLabels, data.evalReviews, data.evalLabels
    charEmbedding, labels = data.charEmbedding, data.labelList
    
    
    train_data = DataGenerator(train_X, train_y)
    eval_data = DataGenerator(eval_X, eval_y)
    pack_data = [train_data, eval_data]
    
    tf.reset_default_graph()
    ## 设置计算图的配置
    session_conf = tf.ConfigProto(allow_soft_placement=True, log_device_placement=False)
    session_conf.gpu_options.allow_growth = True
    session_conf.gpu_options.per_process_gpu_memory_fraction = 0.9  # 配置GPU占用率
    
    sess = tf.Session(config=session_conf)
    
    ## 创建一个实例
    model = CharCNN(config, charEmbedding)
    logger = Logger(sess, config)
    
    trainer = Trainer(sess, model, pack_data, config, logger)
    
    trainer.train_all()

### 不加BN层

main()

**不加BN层的情况下**

- <font size=3 color=red>**Precision: 0.81009, Recall: 0.85063**</font>

### 全连接层加上BN层

main()

**训练不稳定**

### 卷积层加上BN

main()

**训练不稳定，模型变化比较频繁**

## 使用Yelps数据集训练

def main():
    # 实例化配置参数对象
    ## 执行训练数据文件名
    path = "../data/yelps/yelps_test.csv"
    config = Config(path)
    config["summary_dir"] = "../model/charCNN/yelps/summary"
    config["checkpoint_dir"] = "../model/charCNN/yelps/checkpoint"
    
    
    create_dirs([config['summary_dir'], config["checkpoint_dir"]])
    
    data = Dataset(config)
    ## 生成训练集数据，第一个参数表示wordembedding文件所在的文件夹
    data.dataGen("../data/yelps/", prefix="yelps")
    train_X, train_y, eval_X, eval_y = data.trainReviews, data.trainLabels, data.evalReviews, data.evalLabels
    charEmbedding, labels = data.charEmbedding, data.labelList
    
    
    train_data = DataGenerator(train_X, train_y)
    eval_data = DataGenerator(eval_X, eval_y)
    pack_data = [train_data, eval_data]
    
    tf.reset_default_graph()
    ## 设置计算图的配置
    session_conf = tf.ConfigProto(allow_soft_placement=True, log_device_placement=False)
    session_conf.gpu_options.allow_growth = True
    session_conf.gpu_options.per_process_gpu_memory_fraction = 0.9  # 配置GPU占用率
    
    sess = tf.Session(config=session_conf)
    
    ## 创建一个实例
    model = CharCNN(config, charEmbedding)
    logger = Logger(sess, config)
    
    trainer = Trainer(sess, model, pack_data, config, logger)
    
    trainer.train_all()

### 不加BN层

main()
